関数を別のファイルに分けた
