+++ app/javascript/application.jsx
@@ -1,5 +1,6 @@
 import Mount from "./Mount";
 import MountComponents from "./MountComponents";
+import { getDivIds, getMaxId } from "./divUtil"
 
 import Hello from "./components/Hello";
 import Hey from "./components/Hey";
@@ -10,28 +11,7 @@ Mount(Hey, "hey");
 
 // div要素のNodeListを取得
 const divElements = document.querySelectorAll('div');
-
-// NodeListからdivのidsを取得している
-const getDivIds = (divElements) => {
-  const divIds = [];
-  for (let i = 0; i < divElements.length; i++) {
-    if (divElements[i].id) {
-      divIds.push(divElements[i].id);
-    }
-  }
-  return divIds;
-}
-
-// 特定の単語が先頭にある、divのIDの末尾の数字の中で、最大値を取得する
-const getMaxId = (prefix, ids) => {
-  const filteredIds = ids.filter(function(id) {
-    return id.startsWith(prefix);
-  }).map(function(id) {
-    return parseInt(id.replace(prefix, ''));
-  });
-  return Math.max.apply(Math, filteredIds);
-}
-
+// div要素のidsを取得
 const divIds = getDivIds(divElements);
 
 MountComponents(Hola, "hola", getMaxId('hola', divIds));
